Backend/requirements.txt
asgiref==3.10.0
boto3==1.40.45
botocore==1.40.45
certifi==2025.10.5
charset-normalizer==3.4.3
cloudinary==1.44.1
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
django-rest-knox==5.0.2
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
googlemaps==4.10.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
packaging==25.0
pillow==11.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
